fix(textobjects): Disable quote pairs text objects for mini.ai
The quotes `"`, `'` and "`" seem to get confused more on complex
situations than the built-in text objects. E.g.

```bash
ybb="$(git config --get-color "" "yellow| black bold")"
```

Mini.ai is unable to find any text object in the sorrounding
quotes as it seems to match the quote pairs from left to right
which means the area from `yellow` to `bold` is not inside a quote.

That line is tricky even for the built-in text objects but overall
they seem to perform more reasonable.

This disables the custom text objects for the 3 type of quotes in
favor of the builtins.