fix(mason): Fix versions of mason and mason-lspconfig
This should be updated when lsp commands are fixed.

It seems that nvim-lspconfig is preparing to support those commands in
nvim 0.11.2. That should be a good point to update the lsp-settings
module to change to `vim.lsp.enable`.

Only outstanding part is addapt the `lsp_handler` to not call lspconfig.
Maybe a new function that works similar but just enables the lsp could
be use instead.